Updated transfer scripts
Updated transfer scripts to allow for different local and remote drive letters.  Sometimes remote machines may not have the same drives or drive letters as the orchestration server so this adjustment to the transfer scripts allows for transferring when the paths don't match exactly between the local and remote server(s).